+++ utils/controller.py
@@ -1,0 +1,3 @@
+def get_user_info(users_data:list)->None:
+    for user in users_data:
+        print(f'Twoj znajomy {user["name"]} z miejscowosci {user["location"]} opublikowal {user["posts"]} postow.')
